pyproject.toml
[project]
name = "disitech-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-phonenumber-field[phonenumbers]>=8.4.0",
    "python-environ>=0.4.54",
]
